refactor(lib::library_db): move "TrackDB" to own file
